+++ Sources/KeyboardKit/Device/InterfaceOrientation.swift
@@ -46,18 +46,18 @@ public extension InterfaceOrientation {
     /// Whether or not the orientation is landscape.
     var isLandscape: Bool {
         switch self {
-        case .portrait, .portraitUpsideDown: return false
-        case .landscape, .landscapeLeft, .landscapeRight: return true
-        case .unknown: return false
+        case .portrait, .portraitUpsideDown: false
+        case .landscape, .landscapeLeft, .landscapeRight: true
+        case .unknown: false
         }
     }
 
     /// Whether or not the orientation is portrait.
     var isPortrait: Bool {
         switch self {
-        case .portrait, .portraitUpsideDown: return true
-        case .landscape, .landscapeLeft, .landscapeRight: return false
-        case .unknown: return false
+        case .portrait, .portraitUpsideDown: true
+        case .landscape, .landscapeLeft, .landscapeRight: false
+        case .unknown: false
         }
     }
 }