+++ RELEASE_NOTES.md
@@ -22,7 +22,8 @@ Due to complexities with the style provider concept, some views still use a styl
 ### ✨ Features
 
 * `Autocomplete.ToolbarStyle` has support for more styling.
-* `Autocomplete.ToolbarItemStyle` has support for more explicit styling.
+* `Autocomplete.ToolbarItemStyle` has support for more styling.
+* `DeviceType` defines a new `.vision` device type and supports visionOS.
 * `Keyboard.Background` now implements `View` and can be used right away.
 * `Keyboard.Toolbar` is a new view that applies a minimum height to its content.
 * `Keyboard.ToolbarStyle` is a new style that can be applied with a view modifier.